ps9: Refactor ExtensionActionViewController
This patch refactors ExtensionActionViewController to increase
readability and reduce duplicated logics.

It introduces ExtensionActionViewController::TriggerSidebarWithUrl which
has sidebar dedicated logics. ExecuteAction chooses Sidebar or Popup
depending on its extension_action. Also, By introducing
PressButtonWithSlideOutIfEnabled and RaiseButton, we can remove copy and
pasted codes from ExtensionActionViewController. We don't need to check
several times whether we are using sidebar or not. And it removes
duplicated codes to check whether we are using redesigned extension
action or not.

Another thing is, If we were showing a popup already, then we should
treat the action to open the same one as a desire to close it (like
clicking a menu button that was already open).

This patch adds is_showing_popup_ to close already opened popup
properly.  Just relying on popup_host_ is not enough, because it can be
set as nullptr via OnExtensionHostDestroyed, before executing
TriggerPopupWithUrl. It causes reopening popup instead of closing it.

Still ExtensionActionViewController::is_showing_popup() is needed for
above purpose. And It is essential to test popup functionality in
BrowserActionsBarBrowserTest. If we change is_showing_popup() to
is_showing_content(), we cannot test popup functionality properly in
browser test.