ps9: Avoid adding is_showing_sidebar_ in ExtensionActionViewController
It was added at 2c464d6af45a ps9: Refactor ExtensionActionViewController

However, it makes a regression which blocks re-open a popup which was closed
by losing focus. To re-open this popup, we need to click the button twice.

We need to modify ExtensionViewHost's behavior to fix this issue.  However, it
is a way out of this patch and also this behavior is reproduced at
8d70289c9413b46631bb04e70982cb130c18fcfa already, so it should be handled in
seperated issue.